infra/pyproject.toml
[project]
name = "sagemaker-exploration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib==2.180.0",
    "constructs>=10.0.0,<11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
